[connection/switch.py] set the terminal window size of the child tty
If the output of some command runs through a pager, pexpect will
hang on expecting the prompt.

This is observed in case on Onyx OS 'show docker ps' when there are
several dockers running, so the output is done through the pager.

This patch sets a big enough terminal size to avoid that issue,
since I did not find a cli command to force disable pager on Onyx OS.

Signed-off-by: Stepan Blyschak <stepanb@mellanox.com>