[advanced-reboot] use fast IO thread also for fast-reboot
Original approach to measure downtime is slow in detecting state
transitions (data plane down -> up). This is demonstrated with the below
log:

2020-08-19 14:19:14 : Send    10 Received    10 ping DUT
2020-08-19 14:19:15 : Send     1 Received     1 arp ping
2020-08-19 14:19:16 : Send   100 Received   100 servers->t1
2020-08-19 14:19:20 : Send   500 Received   500 t1->servers
2020-08-19 14:19:20 : Data plane state transition from down to up (500)

As seen for a single iteration it takes 5 sec to detect that data plane
transitioned to an up state. This affects the fast-reboot results and
makes them vary between sequential runs up to +-5 sec.
There is already a better approach although much more complex which is
used in measuring warm reboot losses. This approach utilizes another
thread for fast IO. Fast IO thread allows to detect an order of
millisecond packet losses and gives a pretty reproducible results
without a significant deviation like in the original approach.

This change introduces a parameter for PTF script wether to use the
original approach for both fast/warm reboot or to use the fast IO
thread. A default one is to use the fast IO thread.
Also, extract_no_cpu_replies was removed from the code, as the fast IO
approach will anyway detect all the lossess during fast reboot and so,
if there were losses right after CPU went down they will be counted and
reflected in the test summary, so no need for additional check.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>